+++ src/axom/inlet/Field.cpp
@@ -305,10 +305,8 @@ bool Field::isUserProvided() const
   {
     auto status = static_cast<ReaderResult>(
       static_cast<int>(m_sidreGroup->getView("retrieval_status")->getData()));
-    if(status != ReaderResult::Success)
-    {
-      return false;
-    }
+    // Even if it wasn't of the right type, we still say that it was user-provided
+    return (status != ReaderResult::NotFound);
   }
   return exists();
 }
@@ -438,10 +436,10 @@ Field& Field::registerVerifier(std::function<bool(const Field&)> lambda)
   return *this;
 }
 
-bool Field::verify() const
+bool Field::verify(std::vector<VerificationError>* errors) const
 {
   // If this field was required, make sure something was defined in it
-  if(!verifyRequired(*m_sidreGroup, m_sidreGroup->hasView("value"), "Field"))
+  if(!verifyRequired(*m_sidreGroup, m_sidreGroup->hasView("value"), "Field", errors))
   {
     return false;
   }
@@ -453,7 +451,7 @@ bool Field::verify() const
       "[Inlet] Value did not meet range/valid "
       "value(s) constraints: {0}",
       m_sidreGroup->getPathName());
-    SLIC_WARNING(msg);
+    INLET_VERIFICATION_WARNING(m_sidreGroup->getPathName(), msg, errors);
     return false;
   }
 
@@ -465,15 +463,17 @@ bool Field::verify() const
       "[Inlet] Default value did not meet range/valid "
       "value(s) constraints: {0}",
       m_sidreGroup->getPathName());
-    SLIC_WARNING(msg);
+    INLET_VERIFICATION_WARNING(m_sidreGroup->getPathName(), msg, errors);
     return false;
   }
 
   // Lambda verification step
   if(m_verifier && !m_verifier(*this))
   {
-    SLIC_WARNING(fmt::format("[Inlet] Field failed lambda verification: {0}",
-                             m_sidreGroup->getPathName()));
+    const std::string msg =
+      fmt::format("[Inlet] Field failed lambda verification: {0}",
+                  m_sidreGroup->getPathName());
+    INLET_VERIFICATION_WARNING(m_sidreGroup->getPathName(), msg, errors);
     return false;
   }
   return true;
@@ -553,12 +553,12 @@ std::string Field::name() const
                       m_sidreGroup->getPathName());
 }
 
-bool AggregateField::verify() const
+bool AggregateField::verify(std::vector<VerificationError>* errors) const
 {
   return std::all_of(
     m_fields.begin(),
     m_fields.end(),
-    [](const VerifiableScalar& field) { return field.verify(); });
+    [&errors](const VerifiableScalar& field) { return field.verify(errors); });
 }
 
 AggregateField& AggregateField::required(bool isRequired)