+++ src/axom/inlet/Verifiable.hpp
@@ -16,6 +16,8 @@
 
 #include <functional>
 
+#include "axom/inlet/inlet_utils.hpp"
+
 namespace axom
 {
 namespace inlet
@@ -83,9 +85,15 @@ public:
   /*!
    *****************************************************************************
    * \brief Verifies the object to make sure it satisfies the imposed requirements
+   * \param [in] errors An optional vector of errors to append to in the case
+   * of verification failure
+   * 
+   * Ownership is not taken of @a errors, the raw pointer is only used for its
+   * optional reference semantics, as opposed to something like
+   * std::optional<std::reference_wrapper<T>>
    *****************************************************************************
   */
-  virtual bool verify() const = 0;
+  virtual bool verify(std::vector<VerificationError>* errors = nullptr) const = 0;
 };
 
 /*!
@@ -112,19 +120,8 @@ public:
 
   // Should not be reassignable
   AggregateVerifiable& operator=(const AggregateVerifiable&) = delete;
-  /*!
-   *****************************************************************************
-   * \brief Set the required status of this object.
-   *
-   * Set whether this object is required, or not, to be in the input file.
-   * The default behavior is to not be required.
-   *
-   * \param [in] isRequired Boolean value of whether object is required
-   *
-   * \return Reference to calling object, for chaining
-   *****************************************************************************
-   */
-  AggregateVerifiable& required(bool isRequired = true)
+
+  AggregateVerifiable& required(bool isRequired = true) override
   {
     for(auto& verifiable : m_verifiables)
     {
@@ -133,17 +130,7 @@ public:
     return *this;
   }
 
-  /*!
-   *****************************************************************************
-   * \brief Return the required status.
-   *
-   * Return that this object is required, or not, to be in the input file.
-   * The default behavior is to not be required.
-   *
-   * \return Boolean value of whether this object is required
-   *****************************************************************************
-   */
-  bool isRequired() const
+  bool isRequired() const override
   {
     return std::any_of(
       m_verifiables.begin(),
@@ -151,15 +138,8 @@ public:
       [](const BaseVerifiable& verifiable) { return verifiable.isRequired(); });
   }
 
-  /*!
-   *****************************************************************************
-   * \brief Registers the function object that will verify this object's contents
-   * during the verification stage.
-   * 
-   * \param [in] The function object.
-   *****************************************************************************
-  */
-  AggregateVerifiable& registerVerifier(std::function<bool(const BaseType&)> lambda)
+  AggregateVerifiable& registerVerifier(
+    std::function<bool(const BaseType&)> lambda) override
   {
     for(auto& verifiable : m_verifiables)
     {
@@ -168,17 +148,13 @@ public:
     return *this;
   }
 
-  /*!
-   *****************************************************************************
-   * \brief Verifies the object to make sure it satisfies the imposed requirements
-   *****************************************************************************
-  */
-  bool verify() const
+  bool verify(std::vector<VerificationError>* errors = nullptr) const override
   {
-    return std::all_of(
-      m_verifiables.begin(),
-      m_verifiables.end(),
-      [](const BaseVerifiable& verifiable) { return verifiable.verify(); });
+    return std::all_of(m_verifiables.begin(),
+                       m_verifiables.end(),
+                       [&errors](const BaseVerifiable& verifiable) {
+                         return verifiable.verify(errors);
+                       });
   }
 
 private: